src/sAgent/quantum_layer/requirements.txt

qiskit>=1.0.0
qiskit-ibm-runtime>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
websockets>=12.0
psutil>=5.9.0
